fix the compatibility problem

+++ dbms/src/Storages/DeltaMerge/tests/MultiSegmentTestUtil.h
@@ -26,7 +26,6 @@
 #include <Storages/StorageDeltaMerge.h>
 #include <Storages/tests/TiFlashStorageTestBasic.h>
 #include <TestUtils/TiFlashTestBasic.h>
-#include <gtest/gtest.h>
 
 #include <cstdint>
 #include <vector>
@@ -106,21 +105,21 @@ public:
         }
         {
             // Check there is only one segment
-            ASSERT_EQ(store->segments.size(), static_cast<unsigned long>(1));
+            ASSERT_EQ(store->segments.size(), 1);
             const auto & [_key, seg] = *store->segments.begin();
             (void)_key;
             ASSERT_EQ(seg->getDelta()->getRows(), n_avg_rows_per_segment * 4);
-            ASSERT_EQ(seg->getStable()->getRows(), static_cast<unsigned long>(0));
+            ASSERT_EQ(seg->getStable()->getRows(), 0);
 
             // Split the segment, now we have two segments with n_rows_per_segment * 2 rows per segment.
             forceForegroundSplit(0);
-            ASSERT_EQ(store->segments.size(), static_cast<unsigned long>(2));
+            ASSERT_EQ(store->segments.size(), 2);
         }
         {
             // Split the 2 segments again.
             forceForegroundSplit(1);
             forceForegroundSplit(0);
-            ASSERT_EQ(store->segments.size(), static_cast<unsigned long>(4));
+            ASSERT_EQ(store->segments.size(), 4);
         }
         {
             std::shared_lock lock(store->read_write_mutex);
@@ -136,7 +135,7 @@ public:
                 ASSERT_GT(expected_stable_rows[i], 0); // We don't check the exact rows of each segment.
                 total_stable_rows += expected_stable_rows[i];
             }
-            ASSERT_EQ(total_stable_rows, static_cast<int>(4 * n_avg_rows_per_segment));
+            ASSERT_EQ(total_stable_rows, 4 * n_avg_rows_per_segment);
         }
         verifyExpectedRowsForAllSegments();
     }
